requirements.txt
Flask==1.1.1
flask-restful==0.3.6
mysql-connector==2.2.9